pyproject.toml
[project]
name = "openrelik-agents-adk"
version = "0.1.0"
description = "Experimental support for agents"
authors = [{ name = "Johan Berggren", email = "jberggren@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["google-adk>=1.5.0", "openrelik-api-client>=0.2.4"]

[dependency-groups]
dev = ["ruff>=0.11.12"]
